Moved files to shared src folder
