--- {{cookiecutter.repo_name}}/problem-templates/cv/src/{{cookiecutter.src_package_name}}/__init__.py
@@ -1,7 +1,0 @@
-"""This package contains modules pertaining to differing parts
-of the end-to-end workflow, excluding the source for serving
-the model through a REST API."""
-
-from . import data_prep
-from . import modeling
-from . import general_utils
