--- {{cookiecutter.repo_name}}/problem-templates/cv/src/{{cookiecutter.src_package_name}}/data_prep/__init__.py
@@ -1,5 +1,0 @@
-"""This `data_prep` module includes module(s) which contains
-utilities for processing, cleaning, or loading data."""
-
-from . import datasets
-from . import transforms
